+++ packages/extensions/dataflow/src/lnode/dataflow-visualisation.vue
@@ -13,6 +13,17 @@
 			:activeLNodeId="activeOutputLNodeId"
 			@change="onActiveOutputLNodeIdChange"
 		/>
+
+		<button :disabled="!activeInputLNode || !activeOutputLNode" class="btn" @click="showModal">
+			+
+		</button>
+
+		<DataflowCreationForm
+			v-if="activeInputLNode && activeOutputLNode"
+			v-model:isOpen="isCreationDialogOpen"
+			:sourceLNode="activeInputLNode"
+			:destinationLNode="activeOutputLNode"
+		/>
 	</div>
 </template>
 
@@ -21,6 +32,7 @@ import LNodeElement from './lnode-element.vue'
 import type { LNodeSDK } from './lnode-database'
 import { onMounted, ref, watch } from 'vue'
 import type { LNode } from '@/lnode/lnode'
+import DataflowCreationForm from './dataflow-creation-form.vue'
 
 const props = defineProps<{
 	lnodeSDK: LNodeSDK | undefined
@@ -30,6 +42,9 @@ const lNodes = ref<LNode[]>([])
 
 const activeInputLNodeId = ref<string | undefined>()
 const activeOutputLNodeId = ref<string | undefined>()
+const activeInputLNode = ref<LNode | null>(null)
+const activeOutputLNode = ref<LNode | null>(null)
+const isCreationDialogOpen = ref(false)
 
 onMounted(initLnode)
 watch(() => props.lnodeSDK, initLnode)
@@ -43,9 +58,19 @@ async function initLnode() {
 
 function onActiveInputLNodeIdChange(newLNodeId?: string) {
 	activeInputLNodeId.value = newLNodeId
+	activeInputLNode.value = getActiveLNodeById(newLNodeId)
 }
 function onActiveOutputLNodeIdChange(newLNodeId?: string) {
 	activeOutputLNodeId.value = newLNodeId
+	activeOutputLNode.value = getActiveLNodeById(newLNodeId)
+}
+
+function getActiveLNodeById(id: string | undefined) {
+	return lNodes.value.find((ln) => ln.id === id) ?? null
+}
+
+function showModal() {
+	isCreationDialogOpen.value = true
 }
 </script>
